pip_requirements.txt
astroid==1.4.7
google-api-python-client==1.5.1
httplib2==0.9.2
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.0
oauth2client==2.2.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pylint==1.6.4
python-dateutil==2.5.3
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
wrapt==1.10.8
wsgiref==0.1.2
